[Gold III] Title: 나머지 합, Time: 136 ms, Memory: 5936 KB -BaekjoonHub

+++ 백준/Gold/10986. 나머지 합/나머지 합.cc
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-long long number[1000001];
+int number[1000001];
 long long remainders[1001];
 
 int main(void)
